% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_overlap_info.R
\name{add_overlap_info}
\alias{add_overlap_info}
\title{Add info about how spots overlap among capture areas}
\usage{
add_overlap_info(spe, metric_name)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} with colData columns \code{exclude_overlapping},
\code{array_row_transformed}, and \code{array_col_transformed}}

\item{metric_name}{character(1) in \code{colnames(colData(spe))}, where
spots belonging to the capture area with highest average value for the metric
take precedence over other spots}
}
\value{
A \code{SpatialExperiment} object with additional \code{colData}
columns \code{spe$exclude_overlapping} and \code{spe$overlap_capture_area}
}
\description{
Given a \code{SpatialExperiment} and column name in its \code{colData},
return a modified copy of the \code{SpatialExperiment} with additional \code{colData}
columns: \code{spe$exclude_overlapping} and \code{spe$overlap_capture_area}
}
\details{
\code{spe$exclude_overlapping} is TRUE for spots with a higher-quality overlapping
capture area and FALSE otherwise. \code{spot_plot} only displays FALSE spots to
prevent overplotting in regions of overlap. \code{spe$overlap_capture_area} gives the
name of the highest-quality overlapping capture area for spots in region of overlap.
Otherwise, it holds the value "none".
}
\author{
Nicholas J. Eagles
}
