% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spe.R
\name{build_spe}
\alias{build_spe}
\title{Build stitched \code{SpatialExperiment}}
\usage{
build_spe(sample_info, coords_dir, metric_name, count_type = "sparse")
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, and \code{spaceranger_dir}}

\item{coords_dir}{A \code{character(1)} vector giving the directory
containing sample directories each with \code{tissue_positions.csv},
\code{scalefactors_json.json}, and \code{tissue_lowres_image.png} files
produced from refinement with Samui}

\item{metric_name}{character(1) in \code{colnames(colData(spe))}, where
spots belonging to the capture area with highest average value for the metric
take precedence over other spots}

\item{count_type}{A \code{character(1)} vector passed to \code{type} from
\code{SpatialExperiment::read10xVisium}, defaulting to "sparse"}
}
\value{
A \code{SpatialExperiment} object with one sample per group specified
in \code{sample_info} using transformed pixel and array coordinates (including
in the \code{spatialCoords}).
}
\description{
First, read in capture-area-level spaceranger outputs. Then, overwrite
spatial coordinates and images to represent group-level samples using
\code{sample_info$group} (though keep original coordinates in
\code{colData} columns ending in "_original"). Next, add info about
overlaps (via \code{spe$exclude_overlapping} and \code{spe$overlap_key}).
Ultimately, return a \code{SpatialExperiment} ready for visualization or
downstream analysis.
}
\examples{
#   For internal testing
\dontrun{
    sample_info = readr::read_csv('dev/test_data/sample_info.csv')
    coords_dir = 'dev/test_data'
    metric_name = 'sum_umi'
    spe = build_spe(sample_info, coords_dir, metric_name)
}
}
\author{
Nicholas J. Eagles
}
