% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_coord_utils.R
\name{.clean_round}
\alias{.clean_round}
\title{Round to the nearest integer, always rounding up at 0.5}
\usage{
.clean_round(x)
}
\arguments{
\item{x}{\code{numeric()} vector.}
}
\value{
A \code{numeric()} vector rounded to the nearest integer.
}
\description{
This consistent behavior is favorable for our application, where we want to
minimize duplicate mappings of spots to new array coordinates.
}
\author{
Nicholas J. Eagles
}
\keyword{internal}
