% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_coord_utils.R
\name{.validate_array}
\alias{.validate_array}
\title{Check if coordinates are Visium-like}
\usage{
.validate_array(coords)
}
\arguments{
\item{coords}{A \code{tibble} containing 'array_row' and 'array_col' columns
calculated internally by \code{add_array_coords()}}
}
\value{
It returns \code{NULL} if all tests were correct.
}
\description{
Sanity check designed to catch unforeseen bugs: halt if the tibble-like
\code{coords}, expected to contain columns 'array_row' and 'array_col',
represents an invalid Visium array
}
\author{
Nicholas J. Eagles
}
\keyword{internal}
