% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imagej_image.R
\name{prep_imagej_image}
\alias{prep_imagej_image}
\title{Create low-res image and scale factors from high-res ImageJ output image}
\usage{
prep_imagej_image(sample_info, out_dir, lowres_max_size = 1200)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, \code{imagej_image_path}, and \code{spaceranger_dir}.}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output image(s) and scale factors. Provided the parent exists, \code{out_dir}
will be created if necessary.}

\item{lowres_max_size}{An integer(1) vector: the resolution (number of
pixels) of the larger dimension of the output image(s), considered to be "low
resolution".}
}
\description{
\link{DESCRIPTION HERE}
}
\examples{
## TODO: add working examples
args(prep_imagej_image)
}
\author{
Nicholas J. Eagles
}
