% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_imagej_xml.R
\name{read_imagej_xml}
\alias{read_imagej_xml}
\title{Parse transform info from ImageJ XML output}
\usage{
read_imagej_xml(sample_info)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, and \code{imagej_out_path}}
}
\value{
A \code{tibble}: a copy of \code{sample_info} with additional columns
\code{transform_x}, \code{transform_y}, and \code{transform_theta}
}
\description{
Given a \code{tibble} of sample information (\code{sample_info}) with
columns \code{capture_area}, \code{group}, and \code{imagej_out_path},
expected to have one unique path to ImageJ XML output per group, return
a copy of \code{sample_info} with additional columns \code{transform_x},
\code{transform_y}, and \code{transform_theta} representing the rigid affine
transforms needed to be applied to capture areas within each group to
produce the proper relative arrangements.
}
\examples{
#   For internal testing
}
\author{
Nicholas J. Eagles
}
