% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_imagej_inputs.R
\name{rescale_imagej_inputs}
\alias{rescale_imagej_inputs}
\title{Write same-scale hires images for input to ImageJ}
\usage{
rescale_imagej_inputs(sample_info, out_dir)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, and \code{spaceranger_dir}}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output images. Provided the parent exists, \code{out_dir}
will be created if necessary.}
}
\value{
A \code{tibble}: a copy of \code{sample_info} with additional columns
\code{intra_group_scalar} and \code{group_hires_scalef}
}
\description{
Given a \code{tibble} of sample information (\code{sample_info}) with
columns \code{capture_area}, \code{group}, and \code{spaceranger_dir},
Write new high-resolution images for use as input to ImageJ. Particularly
when capture areas come from different slides, there is a risk of significant
scale differences among Spaceranger's \code{tissue_hires_image.png} images;
that is, the physical distance represented by a pixel from each capture area
may differ nontrivially, leading to a distance-distorted output image, and
inconsistent scaling when later transforming pixel coordinates. This function
writes approximately high-res images whose pixels are of equal physical size
within each \code{group}, then adds \code{intra_group_scalar} and
\code{group_hires_scalef} columns to \code{sample_info}. \code{intra_group_scalar}
gives the scalar by a which a given capture area's hires image and pixel
coordinates must be multiplied to match the scale of other \code{group}
members; \code{group_hires_scalef} gives the new \code{tissue_hires_scalef}
(as from Spaceranger's \code{scalefactors_json.json} file) appropriate for
every capture area from the group.
}
\examples{
\dontrun{
#   For internal testing
sample_info <- readr::read_csv("dev/test_data/sample_info.csv")
prep_imagej_coords(sample_info, tempdir())
}

## TODO: add working examples
args(rescale_imagej_inputs)
}
\author{
Nicholas J. Eagles
}
