% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spe_small}
\alias{spe_small}
\title{Small \code{SpatialExperiment} of stitched capture areas}
\format{
An object of class \code{SpatialExperiment} with 4 rows and 9548 columns.
}
\usage{
spe_small
}
\description{
A \code{SpatialExperiment} with 4 rows and 9548 columns consisting of two
stitched-together capture areas in the human lateral septum
}
\keyword{datasets}
