% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_gene.R
\name{spot_plot_z_score}
\alias{spot_plot_z_score}
\title{Spatial plot of multiple genes combined by averaging Z scores}
\usage{
spot_plot_z_score(
  spe,
  genes,
  sample_id,
  assayname = "logcounts",
  minCount = 0,
  ...
)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} with colData column \code{exclude_overlapping},
passed to \code{spatialLIBD::vis_gene} or \code{spatialLIBD::vis_clus}}

\item{genes}{character() of gene names to plot in combination, expected to be
in \code{rownames(spe)}}

\item{sample_id}{character(1) passed to \code{sampleid} in
\code{spatialLIBD::vis_gene} or \code{spatialLIBD::vis_clus}. Assumed to be a
donor, possibly consisting of several capture areas to plot at once}

\item{assayname}{character(1) passed to \code{spatialLIBD::vis_gene}}

\item{minCount}{numeric(1) passed to passed to \code{spatialLIBD::vis_gene}}

\item{...}{Parameters accepted by \code{spot_plot}, excluding
\code{is_discrete} or \code{var_name}, which are handled internally}
}
\value{
A \code{ggplot} object containing a "spot plot" of the specified
sample and genes
}
\description{
This function wraps around \code{spot_plot}, plotting a summary of expression
across multiple genes in a single spatial plot. A Z-score of each gene's
expression across all spots is taken (computed for each spot). To produce a
single value for each spot, Z-scores are averaged across all \code{genes}.
This value is plotted spatially with \code{spot_plot}.
}
\examples{

#   Grab an example SpatialExperiment and suppose all of its spots should be
#   plotted (for spatialNAc, 'exclude_overlapping' will only have genuinely
#   overlapping spots be TRUE)
spe <- spatialLIBD::fetch_data(type = "spatialDLPFC_Visium_example_subset")
spe$exclude_overlapping <- FALSE

white_matter_genes = c(
    "ENSG00000197971", "ENSG00000131095", "ENSG00000123560",
    "ENSG00000171885"
)
spot_plot_z_score(
   spe = spe,
   genes = white_matter_genes,
   sample_id = unique(spe)$sample_id[1],
   assayname = 'logcounts'
)
}
\author{
Nicholas J. Eagles
}
\concept{Spot plots summarizing expression of multiple genes simultaneously}
