% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_overlap_info.R
\name{add_overlap_info}
\alias{add_overlap_info}
\title{Add info about how spots overlap among capture areas}
\usage{
add_overlap_info(spe, metric_name)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} with colData columns \code{exclude_overlapping},
\code{array_row_transformed}, \code{array_col_transformed}, \code{key}, and
\code{sample_id_original}}

\item{metric_name}{character(1) in \code{colnames(colData(spe))}, where
spots belonging to the capture area with highest average value for the metric
take precedence over other spots}
}
\value{
A \code{SpatialExperiment} object with additional \code{colData}
columns \code{spe$exclude_overlapping} and \code{spe$overlap_capture_area}
}
\description{
Given a \code{SpatialExperiment} and column name in its \code{colData},
return a modified copy of the \code{SpatialExperiment} with additional \code{colData}
columns: \code{spe$exclude_overlapping} and \code{spe$overlap_key}
}
\details{
\code{spe$exclude_overlapping} is TRUE for spots with a higher-quality overlapping
capture area and FALSE otherwise. \code{spot_plot} only displays FALSE spots to
prevent overplotting in regions of overlap. \code{spe$overlap_key} gives comma-
separated strings containing the keys of any overlapping spots, and is the empty
string otherwise.
}
\examples{
#   For internal testing
\dontrun{
    library(HDF5Array)
    spe = loadHDF5SummarizedExperiment('dev/test_data/spe_filtered')
    spe_new = add_overlap_info(spe, metric_name = 'sum_umi')
    p_old = spot_plot(
        spe, sample_id = 'Br8492', var_name = 'sample_id_original',
        is_discrete = TRUE
    )
    p_new = spot_plot(
        spe_new, sample_id = 'Br8492', var_name = 'sample_id_original',
        is_discrete = TRUE
    )
    print(p_old)
    print(p_new)
}
}
\author{
Nicholas J. Eagles
}
