% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spe.R
\name{build_spe}
\alias{build_spe}
\title{Build stitched \code{SpatialExperiment}}
\usage{
build_spe(
  sample_info,
  coords_dir,
  count_type = "sparse",
  reference_gtf = NULL,
  gtf_cols = c("source", "type", "gene_id", "gene_version", "gene_name", "gene_type")
)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area} and
\code{group}}

\item{coords_dir}{A \code{character(1)} vector giving the directory
containing sample directories each with \code{tissue_positions.csv},
\code{scalefactors_json.json}, and \code{tissue_lowres_image.png} files
produced from refinement with \code{prep_imagej_*()} functions}

\item{count_type}{A \code{character(1)} vector passed to \code{type} from
\code{SpatialExperiment::read10xVisium}, defaulting to "sparse"}

\item{reference_gtf}{Passed to \code{\link[spatialLIBD:read10xVisiumWrapper]{spatialLIBD::read10xVisiumWrapper()}}}

\item{gtf_cols}{Passed to \code{\link[spatialLIBD:read10xVisiumWrapper]{spatialLIBD::read10xVisiumWrapper()}}}
}
\value{
A \code{SpatialExperiment} object with one sample per group specified
in \code{sample_info} using transformed pixel and array coordinates (including
in the \code{spatialCoords}).
}
\description{
First, read in capture-area-level spaceranger outputs. Then, overwrite
spatial coordinates and images to represent group-level samples using
\code{sample_info$group} (though keep original coordinates in
\code{colData} columns ending in "_original"). Next, add info about
overlaps (via \code{spe$exclude_overlapping} and \code{spe$overlap_key}).
Ultimately, return a \code{SpatialExperiment} ready for visualization or
downstream analysis.
}
\examples{
########################################################################
#   Prepare sample_info
########################################################################

sample_info = tibble(
    group = "Br2719",
    capture_area = c("V13B23-283_A1", "V13B23-283_C1", "V13B23-283_D1")
)
#   Add 'spaceranger_dir' column
sr_dir = tempdir()
temp = unzip(fetch_data("Visium_LS_spaceranger"), exdir = sr_dir)
sample_info$spaceranger_dir = file.path(
    sr_dir, sample_info$capture_area, 'outs', 'spatial'
)

#   Add ImageJ-output-related columns
imagej_dir = tempdir()
temp = unzip(fetch_data("Visium_LS_ImageJ_out"), exdir = imagej_dir)
sample_info$imagej_xml_path = temp[grep('xml$', temp)]
sample_info$imagej_image_path = temp[grep('png$', temp)]

sample_info = rescale_imagej_inputs(sample_info, out_dir = tempdir())

spe_input_dir = tempdir()
prep_imagej_coords(sample_info, out_dir = spe_input_dir)
prep_imagej_image(sample_info, out_dir = spe_input_dir)

########################################################################
#   Build the SpatialExperiment
########################################################################

#    Retrieve a GTF from GENCODE
bfc <- BiocFileCache()
gtf_cache <- bfcrpath(
    bfc,
    paste0(
        "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/",
        "release_32/gencode.v32.annotation.gtf.gz"
    )
)

spe = build_spe(
    sample_info, coords_dir = spe_input_dir, reference_gtf = gtf_cache
)
spe
}
\author{
Nicholas J. Eagles
}
