% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imagej_coords.R
\name{prep_imagej_coords}
\alias{prep_imagej_coords}
\title{Apply transform info from ImageJ XML output}
\usage{
prep_imagej_coords(sample_info, out_dir)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, and \code{imagej_xml_path}}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output pixel coordinates CSVs. Provided the parent exists, \code{out_dir}
will be created if necessary.}
}
\description{
Given a \code{tibble} of sample information (\code{sample_info}) with
columns \code{capture_area}, \code{group}, and \code{imagej_xml_path},
expected to have one unique path to ImageJ XML output per group, read in
the pixel coordinates from each capture area's \code{tissue_positions.csv}
file from Spaceranger, and transform using the rotation matrix specified
by ImageJ. Write one new \code{tissue_positions.csv} file per group
a copy of \code{sample_info} with additional columns \code{transform_x},
\code{transform_y}, and \code{transform_theta} representing the rigid affine
transforms needed to be applied to capture areas within each group to
produce the proper relative arrangements.
}
\examples{
\dontrun{
#   For internal testing
sample_info <- readr::read_csv("dev/test_data/sample_info.csv")
prep_imagej_coords(sample_info, tempdir())
}

## TODO: add working examples
args(prep_imagej_coords)
}
\author{
Nicholas J. Eagles
}
