% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imagej_coords.R
\name{prep_imagej_coords}
\alias{prep_imagej_coords}
\title{Apply transform info from ImageJ XML output}
\usage{
prep_imagej_coords(sample_info, out_dir)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, \code{imagej_xml_path}, \code{intra_group_scalar}, and
\code{group_hires_scalef}}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output pixel coordinates CSVs. Provided the parent exists, \code{out_dir}
will be created if necessary.}
}
\description{
Given a \code{tibble} of sample information (\code{sample_info}) with
columns \code{capture_area}, \code{group}, and \code{imagej_xml_path},
expected to have one unique path to ImageJ XML output per group, read in
the pixel coordinates from each capture area's \code{tissue_positions.csv}
file from Spaceranger, and transform using the rotation matrix specified
by ImageJ. Write one new \code{tissue_positions.csv} file per group.
}
\examples{
#    Define sample information for the example human brain data
sample_info <- dplyr::tibble(
    group = "Br2719",
    capture_area = c("V13B23-283_A1", "V13B23-283_C1", "V13B23-283_D1")
)
#   Add 'spaceranger_dir' column
sr_dir <- tempdir()
temp <- unzip(
    spatialLIBD::fetch_data("visiumStitched_brain_spaceranger"),
    exdir = sr_dir
)
sample_info$spaceranger_dir <- file.path(
    sr_dir, sample_info$capture_area, "outs", "spatial"
)

#   Add ImageJ-output-related columns
imagej_dir <- tempdir()
temp <- unzip(
    spatialLIBD::fetch_data("visiumStitched_brain_ImageJ_out"),
    exdir = imagej_dir
)
sample_info$imagej_xml_path <- temp[grep("xml$", temp)]
sample_info$imagej_image_path <- temp[grep("png$", temp)]

sample_info <- rescale_imagej_inputs(sample_info, out_dir = tempdir())

spe_input_dir <- tempdir()
prep_imagej_coords(sample_info, out_dir = spe_input_dir)

#    A file of spatial coordinates for the stitched Br2719 was produced
list.files(spe_input_dir)
}
\author{
Nicholas J. Eagles
}
