% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imagej_image.R
\name{prep_imagej_image}
\alias{prep_imagej_image}
\title{Create low-res images and scale factors from high-res ImageJ output images}
\usage{
prep_imagej_image(sample_info, out_dir, lowres_max_size = 1200)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, \code{imagej_image_path}, \code{spaceranger_dir},
\code{intra_group_scalar}, and \code{group_hires_scalef}}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output image(s) and scale factors. Provided the parent exists, \code{out_dir}
will be created if necessary.}

\item{lowres_max_size}{An integer(1) vector: the resolution (number of
pixels) of the larger dimension of the output image(s), considered to be "low
resolution".}
}
\description{
After stitching all groups in \code{sample_info} with ImageJ, images of
various resolutions (pixel dimensions) are left. This function creates copies
of each image whose largest dimension is \code{lowres_max_size} pixels. It
also creates a corresponding \code{scalefactors_json.json} file much like
Spaceranger's. In conjunction with \code{prep_imagej_image()}, this function
prepares for building the \code{SpatialExperiment} with \code{build_spe()}.
}
\examples{
#    Define sample information for the example LS data 
sample_info = dplyr::tibble(
    group = "Br2719",
    capture_area = c("V13B23-283_A1", "V13B23-283_C1", "V13B23-283_D1")
)
#   Add 'spaceranger_dir' column
sr_dir = tempdir()
temp = unzip(
    spatialLIBD::fetch_data("visiumStitched_brain_spaceranger"), exdir = sr_dir
)
sample_info$spaceranger_dir = file.path(
    sr_dir, sample_info$capture_area, 'outs', 'spatial'
)

#   Add ImageJ-output-related columns
imagej_dir = tempdir()
temp = unzip(
    spatialLIBD::fetch_data("visiumStitched_brain_ImageJ_out"), exdir = imagej_dir
)
sample_info$imagej_xml_path = temp[grep('xml$', temp)]
sample_info$imagej_image_path = temp[grep('png$', temp)]

sample_info = rescale_imagej_inputs(sample_info, out_dir = tempdir())

spe_input_dir = tempdir()
prep_imagej_image(
    sample_info, out_dir = spe_input_dir, lowres_max_size = 1000
)

#    A "low resolution" stitched image was produced, which has 1000
#    pixels in its largest dimension
this_image <- imager::load.image(
    file.path(spe_input_dir, "Br2719", 'tissue_lowres_image.png')
)
dim(this_image)
}
\author{
Nicholas J. Eagles
}
