% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_imagej_image.R
\name{prep_imagej_image}
\alias{prep_imagej_image}
\title{Create low-res images and scale factors from high-res ImageJ output images}
\usage{
prep_imagej_image(sample_info, out_dir, lowres_max_size = 1200)
}
\arguments{
\item{sample_info}{A \code{tibble} with columns \code{capture_area},
\code{group}, \code{imagej_image_path}, \code{spaceranger_dir},
\code{intra_group_scalar}, and \code{group_hires_scalef}}

\item{out_dir}{A character(1) vector giving a path to a directory to place
the output image(s) and scale factors. Provided the parent exists, \code{out_dir}
will be created if necessary.}

\item{lowres_max_size}{An integer(1) vector: the resolution (number of
pixels) of the larger dimension of the output image(s), considered to be "low
resolution".}
}
\description{
After stitching all groups in \code{sample_info} with ImageJ, images of
various resolutions (pixel dimensions) are left. This function creates copies
of each image whose largest dimension is \code{lowres_max_size} pixels. It
also creates a corresponding \code{scalefactors_json.json} file much like
Spaceranger's. In conjunction with \code{prep_imagej_image()}, this function
prepares for building the \code{SpatialExperiment} with \code{build_spe()}.
}
\examples{
\dontrun{
#   For internal testing
sample_info <- readr::read_csv("dev/test_data/sample_info.csv")
prep_imagej_image(sample_info, tempdir())
}

## TODO: add working examples
args(prep_imagej_image)
}
\author{
Nicholas J. Eagles
}
